fix: Fix several tests after rebase to pytest PRs
